requirements.txt
asgiref==3.5.2
Django==4.0.6
django-cors-headers==3.13.0
django-filter==1.1.0
django-utils-six==2.0
djangorestframework==3.13.1
djangorestframework-filters==0.11.1
djangorestframework-simplejwt==5.2.0
Markdown==3.4.1
psycopg2==2.9.3
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
whitenoise==6.2.0
